fix: 更新画布交互
